fix: pr fixes
