+++ src/components/Common/ReorderableList/ReorderableList.module.scss
@@ -155,15 +155,3 @@
     z-index: 2;
   }
 }
-
-.visuallyHidden {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border: 0;
-}
